util: move determineSourceRemote to util file
determineSourceRemote() function is being used in more then one place and isn't
specific to mr_create, where it was initialy declared. This patch moves it to
util.go file.
